Extract json module from "snabb config listen"
